# openapi.json
{"openapi":"3.1.0","info":{"title":"Server Intelligence API","description":"Backend API for the Server Intelligence Platform","version":"0.2.0"},"paths":{"/api/metrics/":{"get":{"tags":["Metrics"],"summary":"List Metrics","operationId":"list_metrics_api_metrics__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/metrics/current":{"get":{"tags":["Metrics"],"summary":"Current Metrics","operationId":"current_metrics_api_metrics_current_get","parameters":[{"name":"host","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Host"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/metrics/{metric_name}":{"get":{"tags":["Metrics"],"summary":"Metric History","operationId":"metric_history_api_metrics__metric_name__get","parameters":[{"name":"metric_name","in":"path","required":true,"schema":{"type":"string","title":"Metric Name"}},{"name":"host","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Host"}},{"name":"start","in":"query","required":false,"schema":{"type":"string","default":"-1h","title":"Start"}},{"name":"end","in":"query","required":false,"schema":{"type":"string","default":"now","title":"End"}},{"name":"step","in":"query","required":false,"schema":{"type":"string","default":"30s","title":"Step"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/auth/login":{"post":{"tags":["Authentication"],"summary":"Login","operationId":"login_api_auth_login_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TokenResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/auth/me":{"get":{"tags":["Authentication"],"summary":"Get Me","operationId":"get_me_api_auth_me_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AuthenticatedUser"}}}}},"security":[{"HTTPBearer":[]}]}},"/api/auth/test/viewer":{"get":{"tags":["Authentication"],"summary":"Test Viewer Access","operationId":"test_viewer_access_api_auth_test_viewer_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}},"security":[{"HTTPBearer":[]}]}},"/api/auth/test/operator":{"get":{"tags":["Authentication"],"summary":"Test Operator Access","operationId":"test_operator_access_api_auth_test_operator_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}},"security":[{"HTTPBearer":[]}]}},"/api/auth/test/admin":{"get":{"tags":["Authentication"],"summary":"Test Admin Access","operationId":"test_admin_access_api_auth_test_admin_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}},"security":[{"HTTPBearer":[]}]}},"/api/users":{"get":{"tags":["Users"],"summary":"List Users","operationId":"list_users_api_users_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/UserResponse"},"type":"array","title":"Response List Users Api Users Get"}}}}},"security":[{"HTTPBearer":[]}]},"post":{"tags":["Users"],"summary":"Create User","operationId":"create_user_api_users_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserCreate"}}},"required":true},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"HTTPBearer":[]}]}},"/api/users/{user_id}/role":{"patch":{"tags":["Users"],"summary":"Update User Role","operationId":"update_user_role_api_users__user_id__role_patch","security":[{"HTTPBearer":[]}],"parameters":[{"name":"user_id","in":"path","required":true,"schema":{"type":"integer","title":"User Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserRoleUpdate"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/users/{user_id}/status":{"patch":{"tags":["Users"],"summary":"Update User Status","operationId":"update_user_status_api_users__user_id__status_patch","security":[{"HTTPBearer":[]}],"parameters":[{"name":"user_id","in":"path","required":true,"schema":{"type":"integer","title":"User Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserStatusUpdate"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/settings/ai":{"get":{"tags":["AI Settings"],"summary":"Get Ai Settings","operationId":"get_ai_settings_api_settings_ai_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AISettingResponse"}}}}},"security":[{"HTTPBearer":[]}]},"put":{"tags":["AI Settings"],"summary":"Update Ai Settings","operationId":"update_ai_settings_api_settings_ai_put","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AISettingUpdate"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AISettingResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"HTTPBearer":[]}]}},"/api/settings/ai/key":{"delete":{"tags":["AI Settings"],"summary":"Remove Ai Api Key","operationId":"remove_ai_api_key_api_settings_ai_key_delete","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AISettingResponse"}}}}},"security":[{"HTTPBearer":[]}]}},"/api/assistant/chat":{"post":{"tags":["AI Assistant"],"summary":"Chat With Assistant","operationId":"chat_with_assistant_api_assistant_chat_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ChatRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ChatResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"HTTPBearer":[]}]}},"/api/intelligence/forecast":{"get":{"tags":["Intelligence"],"summary":"Get Forecast","description":"Returns 5m, 15m, 30m, 1h, 3h predictions for CPU, Memory, and Load 1m.\nRespects saved strategy ('persistence' vs 'model') from metadata.","operationId":"get_forecast_api_intelligence_forecast_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}},"security":[{"HTTPBearer":[]}]}},"/api/intelligence/anomaly":{"get":{"tags":["Intelligence"],"summary":"Get Anomaly","description":"Returns current server anomaly score, severity, and anomaly classification.\nEvaluated using trained Isolation Forest pipeline.","operationId":"get_anomaly_api_intelligence_anomaly_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}},"security":[{"HTTPBearer":[]}]}},"/api/alerts":{"post":{"tags":["alerts"],"summary":"Create New Alert","operationId":"create_new_alert_api_alerts_post","requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AlertCreate"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"get":{"tags":["alerts"],"summary":"List Alerts","operationId":"list_alerts_api_alerts_get","parameters":[{"name":"status","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Status"}},{"name":"severity","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Severity"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/alerts/{alert_id}":{"get":{"tags":["alerts"],"summary":"Read Alert","operationId":"read_alert_api_alerts__alert_id__get","parameters":[{"name":"alert_id","in":"path","required":true,"schema":{"type":"integer","title":"Alert Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/alerts/{alert_id}/acknowledge":{"post":{"tags":["alerts"],"summary":"Acknowledge Existing Alert","operationId":"acknowledge_existing_alert_api_alerts__alert_id__acknowledge_post","parameters":[{"name":"alert_id","in":"path","required":true,"schema":{"type":"integer","title":"Alert Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/alerts/{alert_id}/resolve":{"post":{"tags":["alerts"],"summary":"Resolve Existing Alert","operationId":"resolve_existing_alert_api_alerts__alert_id__resolve_post","parameters":[{"name":"alert_id","in":"path","required":true,"schema":{"type":"integer","title":"Alert Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/":{"get":{"summary":"Root","operationId":"root__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/health":{"get":{"summary":"Health","operationId":"health_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}}},"components":{"schemas":{"AISettingResponse":{"properties":{"provider":{"type":"string","title":"Provider"},"model":{"type":"string","title":"Model"},"configured":{"type":"boolean","title":"Configured"},"enabled":{"type":"boolean","title":"Enabled"},"key_preview":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Key Preview"},"updated_by":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Updated By"},"updated_at":{"anyOf":[{"type":"string","format":"date-time"},{"type":"null"}],"title":"Updated At"}},"type":"object","required":["provider","model","configured","enabled"],"title":"AISettingResponse"},"AISettingUpdate":{"properties":{"provider":{"type":"string","title":"Provider","default":"ollama"},"model":{"type":"string","title":"Model","default":"qwen3:1.7b"},"api_key":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Api Key"},"enabled":{"type":"boolean","title":"Enabled","default":true}},"type":"object","title":"AISettingUpdate"},"AlertCreate":{"properties":{"title":{"type":"string","title":"Title"},"message":{"type":"string","title":"Message"},"severity":{"type":"string","title":"Severity","default":"INFO"},"source":{"type":"string","title":"Source","default":"system"},"metric":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Metric"},"server":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Server"}},"type":"object","required":["title","message"],"title":"AlertCreate"},"AuthenticatedUser":{"properties":{"id":{"type":"integer","title":"Id"},"username":{"type":"string","title":"Username"},"role":{"$ref":"#/components/schemas/UserRole"},"is_active":{"type":"boolean","title":"Is Active"}},"type":"object","required":["id","username","role","is_active"],"title":"AuthenticatedUser"},"ChatMessage":{"properties":{"role":{"type":"string","title":"Role","description":"'user' or 'assistant'"},"content":{"type":"string","maxLength":4000,"minLength":1,"title":"Content"}},"type":"object","required":["role","content"],"title":"ChatMessage"},"ChatRequest":{"properties":{"messages":{"items":{"$ref":"#/components/schemas/ChatMessage"},"type":"array","maxItems":20,"minItems":1,"title":"Messages"}},"type":"object","required":["messages"],"title":"ChatRequest"},"ChatResponse":{"properties":{"message":{"type":"string","title":"Message"},"provider":{"type":"string","title":"Provider"},"model":{"type":"string","title":"Model"}},"type":"object","required":["message","provider","model"],"title":"ChatResponse"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"LoginRequest":{"properties":{"username":{"type":"string","maxLength":50,"minLength":3,"title":"Username"},"password":{"type":"string","maxLength":128,"minLength":1,"title":"Password"}},"type":"object","required":["username","password"],"title":"LoginRequest"},"TokenResponse":{"properties":{"access_token":{"type":"string","title":"Access Token"},"token_type":{"type":"string","title":"Token Type","default":"bearer"},"expires_in":{"type":"integer","title":"Expires In"}},"type":"object","required":["access_token","expires_in"],"title":"TokenResponse"},"UserCreate":{"properties":{"username":{"type":"string","maxLength":50,"minLength":3,"pattern":"^[A-Za-z0-9_.-]+$","title":"Username"},"password":{"type":"string","maxLength":128,"minLength":8,"title":"Password"},"role":{"$ref":"#/components/schemas/UserRole","default":"VIEWER"}},"type":"object","required":["username","password"],"title":"UserCreate"},"UserResponse":{"properties":{"id":{"type":"integer","title":"Id"},"username":{"type":"string","title":"Username"},"role":{"$ref":"#/components/schemas/UserRole"},"is_active":{"type":"boolean","title":"Is Active"},"created_at":{"type":"string","format":"date-time","title":"Created At"}},"type":"object","required":["id","username","role","is_active","created_at"],"title":"UserResponse"},"UserRole":{"type":"string","enum":["ADMIN","OPERATOR","VIEWER"],"title":"UserRole"},"UserRoleUpdate":{"properties":{"role":{"$ref":"#/components/schemas/UserRole"}},"type":"object","required":["role"],"title":"UserRoleUpdate"},"UserStatusUpdate":{"properties":{"is_active":{"type":"boolean","title":"Is Active"}},"type":"object","required":["is_active"],"title":"UserStatusUpdate"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"},"input":{"title":"Input"},"ctx":{"type":"object","title":"Context"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}},"securitySchemes":{"HTTPBearer":{"type":"http","scheme":"bearer"}}}}